requirements.txt

torch>=2.1
torchaudio>=2.1
torchvision>=0.16
numpy
tqdm
einops
sentencepiece
soundfile
scipy
Pillow
librosa
